Update Character.java
Did the stuff in the comment. Only issue is that I needed to cast the inputted double to an int because of the way the hitbox class is setup, I did some changing to the hitbox class originally but eventually it would need to be casted to an int anyway for when the rectangles were made.